refactor until compiles
