+++ apps/dog-food/ChatOptions.hs
@@ -3,7 +3,8 @@ module ChatOptions (getChatOpts, ChatOpts (..)) where
 import qualified Data.Attoparsec.ByteString.Char8 as A
 import qualified Data.ByteString.Char8 as B
 import Options.Applicative
-import Simplex.Messaging.Agent.Transmission (SMPServer (..), smpServerP)
+import Simplex.Messaging.Agent.Transmission (smpServerP)
+import Simplex.Messaging.Agent.Types.TransmissionTypes (SMPServer (..))
 
 data ChatOpts = ChatOpts
   { name :: Maybe B.ByteString,
