+++ src/Simplex/Messaging/Agent/Store/SQLite/Util.hs
@@ -17,8 +17,6 @@ module Simplex.Messaging.Agent.Store.SQLite.Util
     updateSndConnWithRcvQueue,
     updateRcvQueueStatus,
     updateSndQueueStatus,
-    insertRcvMsg,
-    insertSndMsg,
   )
 where
 
@@ -33,9 +31,9 @@ import Database.SQLite.Simple.Ok (Ok (Ok))
 import Database.SQLite.Simple.QQ (sql)
 import Database.SQLite.Simple.ToField (ToField (..))
 import Network.Socket (HostName, ServiceName)
-import Simplex.Messaging.Agent.Store
-import Simplex.Messaging.Agent.Store.Types
-import Simplex.Messaging.Agent.Transmission
+import Simplex.Messaging.Agent.Types.ConnTypes
+import Simplex.Messaging.Agent.Types.ErrorTypes
+import Simplex.Messaging.Agent.Types.TransmissionTypes
 import Simplex.Messaging.Protocol as SMP (RecipientId)
 import Text.Read (readMaybe)
 import qualified UnliftIO.Exception as E
@@ -382,50 +380,50 @@ updateSndQueueStatusQuery_ =
     WHERE host = :host AND port = :port AND snd_id = :snd_id;
   |]
 
--- ? rewrite with ExceptT?
-insertRcvMsg :: DB.Connection -> ConnAlias -> AgentMsgId -> AMessage -> IO (Either StoreError ())
-insertRcvMsg dbConn connAlias agentMsgId aMsg =
-  DB.withTransaction dbConn $ do
-    queues <- retrieveConnQueues_ dbConn connAlias
-    case queues of
-      (Just _rcvQ, _) -> do
-        insertMsg_ dbConn connAlias RCV agentMsgId aMsg
-        return $ Right ()
-      (Nothing, Just _sndQ) -> return $ Left SEBadQueueDirection
-      _ -> return $ Left SEBadConn
+-- -- ? rewrite with ExceptT?
+-- insertRcvMsg :: DB.Connection -> ConnAlias -> AgentMsgId -> AMessage -> IO (Either StoreError ())
+-- insertRcvMsg dbConn connAlias agentMsgId aMsg =
+--   DB.withTransaction dbConn $ do
+--     queues <- retrieveConnQueues_ dbConn connAlias
+--     case queues of
+--       (Just _rcvQ, _) -> do
+--         insertMsg_ dbConn connAlias RCV agentMsgId aMsg
+--         return $ Right ()
+--       (Nothing, Just _sndQ) -> return $ Left SEBadQueueDirection
+--       _ -> return $ Left SEBadConn
 
--- ? rewrite with ExceptT?
-insertSndMsg :: DB.Connection -> ConnAlias -> AgentMsgId -> AMessage -> IO (Either StoreError ())
-insertSndMsg dbConn connAlias agentMsgId aMsg =
-  DB.withTransaction dbConn $ do
-    queues <- retrieveConnQueues_ dbConn connAlias
-    case queues of
-      (_, Just _sndQ) -> do
-        insertMsg_ dbConn connAlias SND agentMsgId aMsg
-        return $ Right ()
-      (Just _rcvQ, Nothing) -> return $ Left SEBadQueueDirection
-      _ -> return $ Left SEBadConn
+-- -- ? rewrite with ExceptT?
+-- insertSndMsg :: DB.Connection -> ConnAlias -> AgentMsgId -> AMessage -> IO (Either StoreError ())
+-- insertSndMsg dbConn connAlias agentMsgId aMsg =
+--   DB.withTransaction dbConn $ do
+--     queues <- retrieveConnQueues_ dbConn connAlias
+--     case queues of
+--       (_, Just _sndQ) -> do
+--         insertMsg_ dbConn connAlias SND agentMsgId aMsg
+--         return $ Right ()
+--       (Just _rcvQ, Nothing) -> return $ Left SEBadQueueDirection
+--       _ -> return $ Left SEBadConn
 
--- TODO add parser and serializer for DeliveryStatus? Pass DeliveryStatus?
-insertMsg_ :: DB.Connection -> ConnAlias -> QueueDirection -> AgentMsgId -> AMessage -> IO ()
-insertMsg_ dbConn connAlias qDirection agentMsgId aMsg = do
-  let msg = serializeAgentMessage aMsg
-  ts <- liftIO getCurrentTime
-  DB.executeNamed
-    dbConn
-    insertMsgQuery_
-    [ ":agent_msg_id" := agentMsgId,
-      ":conn_alias" := connAlias,
-      ":timestamp" := ts,
-      ":message" := msg,
-      ":direction" := qDirection
-    ]
+-- -- TODO add parser and serializer for DeliveryStatus? Pass DeliveryStatus?
+-- insertMsg_ :: DB.Connection -> ConnAlias -> QueueDirection -> AgentMsgId -> AMessage -> IO ()
+-- insertMsg_ dbConn connAlias qDirection agentMsgId aMsg = do
+--   let msg = serializeAgentMessage aMsg
+--   ts <- liftIO getCurrentTime
+--   DB.executeNamed
+--     dbConn
+--     insertMsgQuery_
+--     [ ":agent_msg_id" := agentMsgId,
+--       ":conn_alias" := connAlias,
+--       ":timestamp" := ts,
+--       ":message" := msg,
+--       ":direction" := qDirection
+--     ]
 
-insertMsgQuery_ :: Query
-insertMsgQuery_ =
-  [sql|
-    INSERT INTO messages
-      ( agent_msg_id, conn_alias, timestamp, message, direction, msg_status)
-    VALUES
-      (:agent_msg_id,:conn_alias,:timestamp,:message,:direction,"MDTransmitted");
-  |]
+-- insertMsgQuery_ :: Query
+-- insertMsgQuery_ =
+--   [sql|
+--     INSERT INTO messages
+--       ( agent_msg_id, conn_alias, timestamp, message, direction, msg_status)
+--     VALUES
+--       (:agent_msg_id,:conn_alias,:timestamp,:message,:direction,"MDTransmitted");
+--   |]
