+++ src/Simplex/Messaging/Agent/Types/ConnTypes.hs
@@ -1,0 +1,82 @@
+{-# LANGUAGE DataKinds #-}
+{-# LANGUAGE DuplicateRecordFields #-}
+{-# LANGUAGE GADTs #-}
+{-# LANGUAGE KindSignatures #-}
+{-# LANGUAGE StandaloneDeriving #-}
+{-# OPTIONS_GHC -fno-warn-unticked-promoted-constructors #-}
+
+module Simplex.Messaging.Agent.Types.ConnTypes
+  ( ReceiveQueue (..),
+    SendQueue (..),
+    ConnType (..),
+    Connection (..),
+    SConnType (..),
+    SomeConn (..),
+  )
+where
+
+import Data.Kind (Type)
+import Data.Type.Equality
+import Simplex.Messaging.Agent.Types.TransmissionTypes
+import qualified Simplex.Messaging.Protocol as SMP
+import Simplex.Messaging.Types (RecipientPrivateKey, SenderPrivateKey, SenderPublicKey)
+
+data ReceiveQueue = ReceiveQueue
+  { server :: SMPServer,
+    rcvId :: SMP.RecipientId,
+    connAlias :: ConnAlias,
+    rcvPrivateKey :: RecipientPrivateKey,
+    sndId :: Maybe SMP.SenderId,
+    sndKey :: Maybe SenderPublicKey,
+    decryptKey :: DecryptionKey,
+    verifyKey :: Maybe VerificationKey,
+    status :: QueueStatus
+  }
+  deriving (Eq, Show)
+
+data SendQueue = SendQueue
+  { server :: SMPServer,
+    sndId :: SMP.SenderId,
+    connAlias :: ConnAlias,
+    sndPrivateKey :: SenderPrivateKey,
+    encryptKey :: EncryptionKey,
+    signKey :: SignatureKey,
+    status :: QueueStatus
+  }
+  deriving (Eq, Show)
+
+data ConnType = CSend | CReceive | CDuplex deriving (Eq, Show)
+
+data Connection (d :: ConnType) where
+  ReceiveConnection :: ConnAlias -> ReceiveQueue -> Connection CReceive
+  SendConnection :: ConnAlias -> SendQueue -> Connection CSend
+  DuplexConnection :: ConnAlias -> ReceiveQueue -> SendQueue -> Connection CDuplex
+
+deriving instance Show (Connection d)
+
+deriving instance Eq (Connection d)
+
+data SConnType :: ConnType -> Type where
+  SCReceive :: SConnType CReceive
+  SCSend :: SConnType CSend
+  SCDuplex :: SConnType CDuplex
+
+deriving instance Eq (SConnType d)
+
+deriving instance Show (SConnType d)
+
+instance TestEquality SConnType where
+  testEquality SCReceive SCReceive = Just Refl
+  testEquality SCSend SCSend = Just Refl
+  testEquality SCDuplex SCDuplex = Just Refl
+  testEquality _ _ = Nothing
+
+data SomeConn where
+  SomeConn :: SConnType d -> Connection d -> SomeConn
+
+instance Eq SomeConn where
+  SomeConn d c == SomeConn d' c' = case testEquality d d' of
+    Just Refl -> c == c'
+    _ -> False
+
+deriving instance Show SomeConn
